### 2018-03-26 #610 @caley
@arfon perhaps stuartcampbell, dlagrava, or krother.  They all have Java listed and mathematics backgrounds.  Also jochym who seems to have a relevent math/physics background, but not Java.
