### 2018-03-27 #610 @arfon
:wave: @stuartcampbell, @dlagrava and @jochym - would you be willing to review this submission for JOSS. As this is outside my subject area as an editor I'd like to find two reviewers if possible.
