### 2018-03-27 #610 @stuartcampbell
Yes I am willing to have a look...
