### 2018-03-27 #610 @dlagrava
@arfon I can also give a look to this.
