### 2018-03-27 #610 @arfon
@stuartcampbell, @dlagrava - please head over to #650 to carry out the actual review.
