### 2018-03-28 #650 @dlagrava
@arfon This is actually my first review, I really appreciate the JOSS idea and I am very happy to collaborate.

I have started the review and even though the documentation looks good, when trying to run the code I am not getting anywhere (I let the sample run for 1 hour). How should I proceed forward? Do we contact the author? If so, how?

Thanks a lot.
