### 2018-03-28 #650 @arfon
> I have started the review and even though the documentation looks good, when trying to run the code I am not getting anywhere (I let the sample run for 1 hour). How should I proceed forward? Do we contact the author? If so, how?

Thanks for your efforts thus far @dlagrava. We encourage you to engage the author (@caley) either here in this GitHub issue thread or by opening issues on the software repository (https://github.com/caley/OpenEPSim) and linking to those issues here.
