### 2018-03-28 #650 @dlagrava
@caley Hello, I am reviewing your submission to JOSS. Both the installation from the release and compilation from source using maven work as metioned on your README.md.

However, when I try to run the sample code (esap.json), I don't get any results after 1 hour. This both from the release .jar and generating one with maven. I also see no CPU activity from running the code during this time. I have tried both OpenJDK and OracleJDK 1.8 on my Ubuntu machine. Could you tell me if there are extra requirements on the JDK or so?

Thanks a lot.
