### 2018-03-28 #650 @caley
@dlagrava First off, thanks for taking the time to do the review!  The sample file asep.json should take less than 30 seconds to run (13 seconds on my new laptop), so something is going wrong.

The Oracle JDK should definitely be fine, and you would get a error message from the Java run time if there was a problem.  Have you run the command exactly as in the README.md file?  The program expects to read the input configuration from stdin (so you need the input redirection '... < samples/asep.json ...').  If you leave out the '<' you'll get the behaviour you see - no CPU activity, as the program is actually just waiting for input.

If that doesn't work, can you send me the exact commands that you have run?   In any case, I'll update the README to emphasise that it expects the input from stdin.
