### 2018-03-28 #650 @dlagrava
@caley Thanks for the info, this is what I was doing wrong, silly me. I misinterpreted the < and > as markers for the input. Personally I prefer allowing command line parameters for the main program, but this is totally my own opinion.
