### 2018-03-28 c2a5ff7
Note about input and output redirection
