### 2018-03-29 #650 @dlagrava
@caley I have almost finished the review. I am happy with the documentation of the repo and the code. The only thing from the checklist I see missing on the repo is a section with guidelines on how to contribute to the code, file bugs and ask general questions about it. Could you please add this?
