### 2018-03-30 caley/OpenEPSim#1 @stuartcampbell
Suggestion for readme
Hi, I am currently reviewing your submission for JOSS (https://github.com/openjournals/joss-reviews/issues/650) and had a small suggestion for your readme file.  After you get the output, you show some plots.  It would be nice to give the user a pointer or code, or hints on how to go about plotting the output.
Thank you
Stuart
