### 2018-03-30 #650 @stuartcampbell
I'm also nearly done with the review - I agree with @dlagrava about the missing community aspects.
