### 2018-03-30 caley/OpenEPSim#2 @stuartcampbell
Expand CI use
Hi, It looks like you have an initial travis.yml file.  It would be nice if this was fleshed out to run the compile and tests (which worked beautifully).

Cheers
Stu

Link to the JOSS review: https://github.com/openjournals/joss-reviews/issues/650
