### 2018-04-05 b7249c6
Added Travis CI build status badge for #2
