### 2018-04-05 caley/OpenEPSim#2 @caley
Hi Stu,

Travis CI is already compiling the project and running the tests.  I've added a build status badge to the readme to link to this.  Were you suggesting any more than this?

Cheers,

Caley
