### 2018-04-05 caley/OpenEPSim#2 @stuartcampbell
Yeah this is great.  I forgot that travis picks up the `pom.xml` file automagically - and I couldn't find a successful build on travis.  But the badge helps find it.
