### 2018-04-05 12f2e07
Sample code for producing plots of simulation output for #1
