### 2018-04-05 caley/OpenEPSim#1 @caley
Hi Stuart,

I've added some sample code to produce plots like those in the readme.  The code uses Python, because it's simpler to produce quick plots than with Java.

Cheers,

Caley
