### 2018-04-05 c1ce962
Added 'Contributing' section to readme
