### 2018-04-15 #650 @caley
@arfon - I've put it on Zenodo.  The doi is https://doi.org/10.5281/zenodo.1218513.
